// frontend/src/app/model/ws/ws-client-message.interface.ts
import { ClientMessageType } from './client-message-type.enum';
import ClientMessageTypeEnum = ClientMessageType.ClientMessageTypeEnum;
import { IWsMessage } from './ws-message.interface';

/**
 * Structure of a message as generated by the DRES client.
 */
export interface IWsClientMessage extends IWsMessage {
  runId: string;
  type: ClientMessageTypeEnum;
}